grouping stats functions and tests
